fix(show-me): update clipboard and shell demos with preload (#903)
