+++ static/show-me/clipboard/index.html
@@ -1,16 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <title>Clipboard</title>
-</head>
-<body>
-  <h1>Clipboard Demo</h1>
-  <button id="copy">Copy "Hello from Electron!"</button>
-  <button id="paste">Paste into TextArea</button>
-  <br />
-  <textarea></textarea>
-  <script>
-    require('./renderer.js')
-  </script>
-</body>
+  <head>
+    <title>Clipboard</title>
+  </head>
+  <body>
+    <h1>Clipboard Demo</h1>
+    <button id="copy">Copy "Hello from Electron!"</button>
+    <button id="paste">Paste into TextArea</button>
+    <br />
+    <textarea></textarea>
+  </body>
 </html>
