+++ static/show-me/clipboard/main.js
@@ -4,13 +4,14 @@
 // https://electronjs.org/docs/api/clipboard
 
 const { app, BrowserWindow } = require('electron')
+const path = require('path')
 
 app.whenReady().then(() => {
   const mainWindow = new BrowserWindow({
     height: 600,
     width: 600,
     webPreferences: {
-      nodeIntegration: true
+      preload: path.join(__dirname, 'preload.js')
     }
   })
