--- static/show-me/clipboard/renderer.js
@@ -1,13 +1,0 @@
-const { clipboard } = require('electron')
-
-const copyButton = document.querySelector('#copy')
-const pasteButton = document.querySelector('#paste')
-const textarea = document.querySelector('textarea')
-
-copyButton.onclick = () => {
-  clipboard.writeText('Hello from Electron!')
-}
-
-pasteButton.onclick = () => {
-  textarea.value = clipboard.readText()
-}
