+++ static/show-me/shell/index.html
@@ -1,21 +1,12 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <title>Shell Example</title>
-</head>
-<body>
-  <h1>Shell Example</h1>
-  <button id='open-github'>
-    Open GitHub.com
-  </button>
-  <button id='open-folder'>
-    Show me in my folder
-  </button>
-  <button id='beep'>
-    Play a system beep
-  </button>
-  <script>
-    require('./renderer.js')
-  </script>
-</body>
+  <head>
+    <title>Shell Example</title>
+  </head>
+  <body>
+    <h1>Shell Example</h1>
+    <button id="open-github">Open GitHub.com</button>
+    <button id="open-folder">Show me in my folder</button>
+    <button id="beep">Play a system beep</button>
+  </body>
 </html>
